refactor: Change return type of CreateSpeedTestAsync method to int for consistent status reporting

+++ DataProcessingAPI/Application/Interfaces/ISpeedTestService.cs
@@ -6,7 +6,7 @@ public interface ISpeedTestService
 {
     Task<List<SpeedTestDto>> GetSpeedTestResultsAsync(DateTime? startDate, DateTime? endDate);
     Task<SpeedTestDto?> GetSpeedTestByIdAsync(long id);
-    Task<long> CreateSpeedTestAsync(SpeedTestDto speedTest);
+    Task<int> CreateSpeedTestAsync(SpeedTestDto speedTest);
     Task<int> UpdateSpeedTestAsync(long id, SpeedTestDto speedTest);
     Task<int> DeleteSpeedTestAsync(long id);
 }
